+++ src/nether.cpp
@@ -86,7 +86,7 @@ void nether::generateStructures(chunk* generateIn)
 		const blockID currentOre = netherOreList[i];
 		csize_t& count = roundRandom(generateIn->chunkRandom, rarity[i]);
 		cveci2& depositPosition = generateIn->worldPos + veci2(randIndex(generateIn->chunkRandom, (int)chunkSize.x()), randIndex(generateIn->chunkRandom, (int)chunkSize.y()));
-		for (int j = 0; j < count; j++)
+		for (size_t j = 0; j < count; j++)
 		{
 			addOres(depositPosition, currentOre, rand(generateIn->chunkRandom, maxVeinSize[i]), generateIn->chunkRandom, { blockID::netherrack });
 		}
